cdk/requirements.txt
#-e .

aws-cdk.core
aws-cdk.aws_lambda
aws-cdk.aws_apigatewayv2
aws-cdk.aws_apigatewayv2_integrations
